backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
psycopg2-binary
alembic
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
requests
openai
transformers
torch
pandas
scikit-learn
pypdf
python-dotenv
aiofiles
cors
